Refactor chart helpers to a separate file
