+++ pubspec.yaml
@@ -7,7 +7,8 @@ environment:
 
 dependencies:
   firebase: ^7.2.1
-  chartjs: ^0.5.2
+  chartjs:
+    git: git@github.com:marianamarasoiu/chartjs.dart.git
 
 dev_dependencies:
   build_runner: ^1.6.0
